Allow options anywhere on command line.
Well, mostly.

Previously, core options had to come before the command name, and command
options after:

    sqitch [core options] [command] [command-options]

Now they can be mixed and matched, though the ideal is to put all options
after the command name:

    sqitch [command] [options]

I think this change simplifies things for users who were sometimes confused
when options were ignored or considered invalid (of which I was sometimes
one). To accomplish this change, argument parsing has been revamped a fair bit
as follows:

1. First, Sqitch parses core options from all arguments. Sqitch immediately
   triggers any that require no command, such as `--version`, `--etc-path` and
   `--directory`.

2. Next, it searches the remaining arguments for a valid command. If it finds
   no valid command, it reports invalid commands to STDERR (if there were any)
   and exits with the usage statement.

3. Otherwise, it instantiates the command and executes it as before, passing
   the remaining arguments for the command to parse at it sees fit.

In step 2, To try to be as smart as possible about finding a command, Sqitch
searches for an argument that doesn't look like an option -- that is, that
doesn't begin with `--`. That can be a little tricky when using command
options before the command name. For example:

    sqitch --target localhost deploy

In this case, Sqitch will first try to load `localhost`, because it has no way
to tell whether or not `--target` takes an argument, because it hasn't yet
loaded the command and its option spec. So when `localhost` fails to load,
because it's not a command, it next tries `deploy`. When that works, it
continues as if nothing happened.

Still, this can cause issues when a user happens to use an option argument
that corresponds to a command. For example, say they wanted to deploy to a
target named "engine":

    sqitch --target engine deploy

In this case, Sqitch will successfully load the `engine` command and try to
execute it. Which isn't what the user wants. That's why it's recommended that
options always come after the command -- including core options -- so Sqitch
finds it as quickly as possible and won't be confused:

    sqitch deploy --target engine --verbose

If Sqitch finds no valid command, it will report on all of the strings it
attempted to load. For example:

    sqitch --target localhost nonesuch

Would cause Sqitch to report both `localhost` and `nonesuch` as invalid
commands:

    "localhost" is not a valid command
    "nonesuch" is not a valid command

The exception is options considered to be "legacy", and which will soon be
removed from core and into the commands that require them. These include
`--client` and the `--db-*` options. To preserve compatibility for the likely
huge number of scripts and such people use, Sqitch will skip those options and
their arguments without trying to load them. So this command should continue
to work as expected even after the `-d` option moves to the `deploy` command:

    sqitch -d engine deploy

I believe this change does a reasonable job walking the line between what
people likely expect of Sqitch ("just use my options no matter where I put
them!") and preserving backward compatibility ("My `--db-user` option should
still work before the command even if its argument is a valid command name!").